Work GET /product | PostgreSQL
